Test illumos.
