+++ ext/blobio/blob_test.go
@@ -12,6 +12,7 @@ import (
 	_ "github.com/ncruces/go-sqlite3/embed"
 	"github.com/ncruces/go-sqlite3/ext/array"
 	"github.com/ncruces/go-sqlite3/ext/blobio"
+	_ "github.com/ncruces/go-sqlite3/tests/testcfg"
 	_ "github.com/ncruces/go-sqlite3/vfs/memdb"
 )
