+++ ext/pivot/pivot_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/ncruces/go-sqlite3"
 	_ "github.com/ncruces/go-sqlite3/embed"
 	"github.com/ncruces/go-sqlite3/ext/pivot"
+	_ "github.com/ncruces/go-sqlite3/tests/testcfg"
 )
 
 // https://antonz.org/sqlite-pivot-table/
