+++ vfs/memdb/memdb_test.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/ncruces/go-sqlite3"
 	_ "github.com/ncruces/go-sqlite3/embed"
+	_ "github.com/ncruces/go-sqlite3/tests/testcfg"
 )
 
 //go:embed testdata/wal.db
